requirements.txt
strands-agents[openai]
chromadb>=0.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0